requirements.txt
torch>=2.0.0
transformers>=4.30.0
flask>=2.0.0
accelerate>=0.20.0
llama-cpp-python>=0.2.0